Change export structure
See #24
