+++ src/index.ts
@@ -1,30 +1,6 @@
-import MarkdownIt from 'markdown-it';
-import { rolePlugin } from './roles';
-import { directivesPlugin } from './directives';
-import { blocksPlugin } from './blocks';
-import { mathPlugin } from './math/math';
-import directives from './directives/default';
-import roles from './roles/default';
-import { Directives } from './directives/types';
-import { Roles } from './roles/types';
+import * as plugins from './plugins';
+import { roles } from './roles';
+import { directives } from './directives';
 
-export type Options = {
-  directives: Directives;
-  roles: Roles;
-  math?: boolean;
-};
-
-export const defaultPlugins: Options = { directives, roles, math: true };
-export const defaultOpts: MarkdownIt.Options = { html: false };
-
-export default function MyST(
-  plugins: Options = defaultPlugins,
-  opts: MarkdownIt.Options | undefined = defaultOpts,
-) {
-  const tokenizer = MarkdownIt('commonmark', opts);
-  if (plugins.math) tokenizer.use(mathPlugin);
-  tokenizer.use(blocksPlugin);
-  tokenizer.use(directivesPlugin(plugins.directives));
-  tokenizer.use(rolePlugin(plugins.roles));
-  return tokenizer;
-}
+export { default as MyST } from './myst';
+export { plugins, roles, directives };
