+++ src/myst.ts
@@ -1,0 +1,31 @@
+import MarkdownIt from 'markdown-it';
+import * as plugins from './plugins';
+import { directives } from './directives';
+import { roles } from './roles';
+import { Directives } from './directives/types';
+import { Roles } from './roles/types';
+
+export type Options = {
+  directives: Directives;
+  roles: Roles;
+  math?: boolean;
+  markdownit?: MarkdownIt.Options;
+};
+
+export const defaultOptions: Options = {
+  directives,
+  roles,
+  math: true,
+  markdownit: { html: false },
+};
+
+export default function MyST(
+  opts: Options = defaultOptions,
+) {
+  const tokenizer = MarkdownIt('commonmark', opts.markdownit);
+  if (opts.math) tokenizer.use(plugins.math);
+  tokenizer.use(plugins.blocks);
+  tokenizer.use(plugins.directives(directives));
+  tokenizer.use(plugins.roles(roles));
+  return tokenizer;
+}
