+++ src/plugins.ts
@@ -1,0 +1,4 @@
+export { plugin as blocks } from './blocks';
+export { plugin as math } from './math';
+export { plugin as roles } from './roles';
+export { plugin as directives } from './directives';
